SEC SaaS Metrics5/sec-scrappers/xbrl-scrapper/requirements.txt
async-generator==1.10
attrs==21.4.0
boto3==1.21.22
botocore==1.24.22
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chromedriver-binary==100.0.4896.20.0
cryptography==36.0.1
Cython==0.29.28
h11==0.13.0
idna==3.3
jmespath==1.0.0
lxml==4.5.0
numpy==1.22.3
outcome==1.1.0
pandas==1.4.1
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.1.0
